+++ src/user/views/adsManager/views/advanced/components/campaign/fields/LocationField.tsx
@@ -1,0 +1,17 @@
+import {Box, Divider, Typography} from "@mui/material";
+import {LocationPicker} from "../../../../../../../../components/Location/LocationPicker";
+import React from "react";
+
+export function LocationField() {
+  return (
+    <Box mb={5}>
+      <Divider textAlign="left" sx={{fontSize: "24px", mb: 1, mt: 2}}>
+        Location
+      </Divider>
+      <Typography variant="body2" sx={{mb: 5}}>
+        Select the geographic regions where your ads will be shown.
+      </Typography>
+      <LocationPicker states={true}/>
+    </Box>
+  )
+}
